Add analytics plugin
Given a stylesheet that's run through the first phase, check the
original/base selector of each suffixed class selector and see if there
are other places that use it. A percentage is output based off how many
of these base selectors are completely accounted for, so a 100% would
mean that the completion of running all phases on this stylesheet would
mean all those selectors have been modularized.