Add heuristic plugin
The first half of phase 2 of the squeaky process. In order to prepare
the
stylesheet to be flattened (so CSS modules can be used), references to
the base (leaf) class selectors in the stylesheet need to be isolated
to view files that are only associated with the stylesheet. This is
determined
by traversing the selector hierarchy from the bottom up to determine
where the selectors are used within view files and remove them from all
other files in the codebase