Add flags for feature file correlation
On top of analyzing the dependency tree when determining how namespaced
selectors are used, there is a correlation between the feature a
stylesheet is a part of and the feature a given view file is a part of.

This is especially true for "legacy" folder hierarchies that are still
following similar conventions to how Ruby applications are structured
(i.e. `app/assets/...` as opposed to more component based approaches.

Thus, providing a heuristic function helps determine the feature a file
is a part of (typically RegExps that try to parse out the feature from
file names/directories). These are added onto the whitelist of files
that are "blessed" from having a given namespaced selector be removed
from those files (since it's likely that a namespaced selector should
be used in the feature it was declared in)