Add flatten plugin
The final part of phase 2 of squeaky clean. With the heuristic plugin
scraping the namespaced selectors from files except for the view files
that use it, the stylesheet selectors can now be pulled top level so
they target the specific instances where the namespaced selectors are
present.

Include a basic suite of specs which will be enhanced with edge cases