Extract temporary stylesheet path as an option
PostCSS plugins typically iterate through the CSS ruleset once and make
modifications in that iteration as a part of a map.

In order to handle both the removal of the selector namespaces and
index duplicate base selectors, these require two loops which has one
dependent on the other. Since both the dupe indexing and the namespace
removal is done in bash for simplicity sake, moving the file over in
bash preserves all changes